Fix circular import for ComposerRequirements
